+++ core/src/builtin_widgets/text_style.rs
@@ -2,7 +2,6 @@ use crate::{prelude::*, wrap_render::WrapRender};
 
 /// This widget establishes the text style for painting the text within its
 /// descendants.
-#[derive(Default)]
 pub struct TextStyleWidget {
   pub text_style: TextStyle,
 }
@@ -18,16 +17,23 @@ impl<'c> ComposeChild<'c> for TextStyleWidget {
 
   fn compose_child(this: impl StateWriter<Value = Self>, child: Self::Child) -> Widget<'c> {
     // We need to provide the text style for the children to access.
-    match this.try_into_value() {
+    let (child, provider): (_, Box<dyn Query>) = match this.try_into_value() {
       Ok(this) => {
         let style = this.text_style.clone();
-        WrapRender::combine_child(State::value(this), child).attach_data(Box::new(Queryable(style)))
+        (WrapRender::combine_child(State::value(this), child), Box::new(Queryable(style)))
       }
       Err(this) => {
         let style = this.map_reader(|w| PartData::from_ref(&w.text_style));
-        WrapRender::combine_child(this, child).attach_data(Box::new(style))
+        (WrapRender::combine_child(this, child), Box::new(style))
       }
-    }
+    };
+
+    let ctx = BuildCtx::get_mut();
+    ctx.current_providers.push(provider);
+    child.into_widget().on_build(|id| {
+      let provider = ctx.current_providers.pop().unwrap();
+      id.attach_data(provider, ctx.tree_mut());
+    })
   }
 }